+++ gradle.properties
@@ -3,15 +3,15 @@ org.gradle.jvmargs=-Xmx1G
 
 # Fabric Properties
 # check these on https://fabricmc.net/use
-minecraft_version=21w20a
-yarn_mappings=21w20a+build.19
-loader_version=0.11.3
+minecraft_version=1.18.1
+yarn_mappings=1.18.1+build.2
+loader_version=0.12.11
 
 #Fabric api
-fabric_version=0.34.7+1.17
+fabric_version=0.44.0+1.18
 
 # Mod Properties
-mod_version=1.3.0
+mod_version=1.4.0+1.18
 maven_group=draylar
 archives_base_name=structurized
