+++ src/main/java/draylar/structurized/mixin/StructurePoolAccessor.java
@@ -1,13 +1,13 @@
 package draylar.structurized.mixin;
 
+import java.util.List;
+
 import com.mojang.datafixers.util.Pair;
-import net.minecraft.structure.pool.StructurePool;
-import net.minecraft.structure.pool.StructurePoolElement;
 import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.Mutable;
 import org.spongepowered.asm.mixin.gen.Accessor;
 
-import java.util.List;
+import net.minecraft.structure.pool.StructurePool;
+import net.minecraft.structure.pool.StructurePoolElement;
 
 @Mixin(StructurePool.class)
 public interface StructurePoolAccessor {
@@ -17,7 +17,6 @@ public interface StructurePoolAccessor {
     @Accessor(value = "elementCounts")
     List<Pair<StructurePoolElement, Integer>> getElementCounts();
 
-    @Mutable
     @Accessor(value = "elementCounts")
     void setElementCounts(List<Pair<StructurePoolElement, Integer>> list);
 }
