+++ src/main/resources/structurized.mixins.json
@@ -1,7 +1,7 @@
 {
   "required": true,
   "package": "draylar.structurized.mixin",
-  "compatibilityLevel": "JAVA_8",
+  "compatibilityLevel": "JAVA_17",
   "mixins": [
     "DynamicRegistryManagerMixin",
     "StructurePoolAccessor"
